pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yowling-codecks-bot"
version = "0.1.0"
description = "A simple Telegram bot"
dependencies = [
    "python-telegram-bot",
    "python-dotenv",
    "requests",
    "termcolor==2.4.0",
]
